Blaze_tablet : add wallpaper_info.xml with default wallpaper dimensions
Fixes the wallpaper corruption during the first boot on WUXGA tablet2 :

This issue is occurring due to an OutOfMemoryException.
On first boot wallpaper_info.xml doesn't exist sothe system "guesses"
that the wallpaper should be 1920x1920 (the largest of the 2 screen
dimensions as determined by the window manager).
Later, the system realizes that the correct size is 2820x1920 and
scales the 1920x1920 bitmap to fit. That means that, for a moment,
a 1920x1920 and 2820x1920 bitmap exist on the heap of the SystemUI
process and there isn't enough heap memory.

Change-Id: I1093fc6cec9f012ecf27c8cc9d8566bae34293aa
Signed-off-by: Dima Svetlov <svetlov@ti.com>