recovery: fix recovery mode for blaze tablet
There are two issues that breaks recovery. The first one is that
actually mountpoints specified incorrect, using paths that was
actual for linux kernel 3.0. Since 3.4 paths was changed, so we
just create symlink to keep compatibility with both 3.0 and 3.4.
And the second one is that pixelformat BGRA_8888 is not correctly
supported by OMAP framebuffer, so switch to use RGB565.

Change-Id: Ic56b663fa1a2f906a229c25e160251421656d6fc
Signed-off-by: Andrii Beregovenko <a.beregovenko@ti.com>